refresh vol periodically instead of only by signal
as there are different volumes/mutes for headphones/not,
sometimes the volume will show up as muted when it isnt since
it was refreshed (once) when headphones werent plugged in
and the volume hasnt been changed since headphones where plugged in.
This refreshes once every 5 minutes, just to make sure its still
current.