sticky/window/resolution/streaming functionality
lots of helper scripts and i3-picture-in-picture,
which stickies a window in the bottom right of the screen.

reason for all of this is to enable stream-corner.
'$mod+d m' will grab a link from my clipboard,
prompt me to select a resolution using youtube-dl/rofi,
and then stream from it using mpv, putting the
window in the bottom left of the main screen. the
window remains there no matter what workspace
I switch to.